+++ src/main/java/com/tidepool/api/model/Highlight.java
@@ -4,48 +4,50 @@ import java.io.Serializable;
 
 public class Highlight implements Serializable {
 
-	public int x0;
-	public int y0;
-	public int x1;
-	public int y1;
-	public int height;
-	public int width;
+	public double x0;
+	public double y0;
+	public double x1;
+	public double y1;
+	public double height;
+	public double width;
 	
-	public int getX0() {
+	public double getX0() {
 		return x0;
 	}
-	public void setX0(int x0) {
+	public void setX0(double x0) {
 		this.x0 = x0;
 	}
-	public int getY0() {
+	public double getY0() {
 		return y0;
 	}
-	public void setY0(int y0) {
+	public void setY0(double y0) {
 		this.y0 = y0;
 	}
-	public int getX1() {
+	public double getX1() {
 		return x1;
 	}
-	public void setX1(int x1) {
+	public void setX1(double x1) {
 		this.x1 = x1;
 	}
-	public int getY1() {
+	public double getY1() {
 		return y1;
 	}
-	public void setY1(int y1) {
+	public void setY1(double y1) {
 		this.y1 = y1;
 	}
-	public int getHeight() {
+	public double getHeight() {
 		return height;
 	}
-	public void setHeight(int height) {
+	public void setHeight(double height) {
 		this.height = height;
 	}
-	public int getWidth() {
+	public double getWidth() {
 		return width;
 	}
-	public void setWidth(int width) {
+	public void setWidth(double width) {
 		this.width = width;
 	}
+	
+	
 	
 }
